Adds support for TangoLimaVideo on top of GenericVideoDevice and removing all Qt (and Qub) dependencies
Classes GenericVideoDevice already present in "master" branch are added in this 2.2 branch

  Committer: Vicente Rey Bakaikoa

Adds support for TangoLimaVideo on top of GenericVideoDevice and removing all Qt (and Qub)
dependencies

Classes GenericVideoDevice already present in "master" branch are added in this 2.2 branch

  Committer: Vicente Rey Bakaikoa

missing files after git magic
